complete list and delete api

+++ src/main/java/com/star/eswasthyabackend/service/allergy/impl/AllergicMedicineServiceImpl.java
@@ -7,6 +7,7 @@ import com.star.eswasthyabackend.model.patient.PatientDetails;
 import com.star.eswasthyabackend.repository.AllergicMedicineRepository;
 import com.star.eswasthyabackend.repository.user.patient.PatientDetailsRepository;
 import com.star.eswasthyabackend.service.allergy.AllergicMedicineService;
+import com.star.eswasthyabackend.utility.AuthenticationUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -14,6 +15,7 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -21,20 +23,21 @@ public class AllergicMedicineServiceImpl implements AllergicMedicineService {
 
     private final AllergicMedicineRepository allergicMedicineRepository;
     private final PatientDetailsRepository patientDetailsRepository;
+    private final AuthenticationUtil authenticationUtil;
+
     @Override
     public Integer saveAllergyList(AllergicMedicineRequestDto allergicMedicineRequestDto) {
 
         PatientDetails patientDetails = patientDetailsRepository.findById(allergicMedicineRequestDto.getPatientDetailId())
-                .orElseThrow(()-> new AppException("Patient not found for given id.", HttpStatus.BAD_REQUEST));
+                .orElseThrow(() -> new AppException("Patient not found for given id.", HttpStatus.BAD_REQUEST));
         List<AllergicMedicine> allergicMedicineList = new ArrayList<>();
         allergicMedicineRequestDto.getAllergicMedicineList().forEach(
                 allergyRequest -> {
                     AllergicMedicine allergicMedicine;
-                    if(allergyRequest.getId()!=null){
+                    if (allergyRequest.getId() != null) {
                         allergicMedicine = allergicMedicineRepository.findById(allergyRequest.getId())
-                                .orElseThrow(()-> new AppException("Allergic Medicine not found for given id.", HttpStatus.BAD_REQUEST));
-                    }
-                    else{
+                                .orElseThrow(() -> new AppException("Allergic Medicine not found for given id.", HttpStatus.BAD_REQUEST));
+                    } else {
                         allergicMedicine = new AllergicMedicine();
                     }
                     allergicMedicine.setAllergicMedicineName(allergyRequest.getAllergicMedicineName());
@@ -53,6 +56,10 @@ public class AllergicMedicineServiceImpl implements AllergicMedicineService {
 
     @Override
     public Boolean deleteById(Integer allergicMedicineId) {
+        String role = authenticationUtil.getRole();
+        if (!Objects.equals(role, "DOCTOR")) {
+            throw new AppException("You are not authorized to access this resource", HttpStatus.BAD_REQUEST);
+        }
         allergicMedicineRepository.deleteById(allergicMedicineId);
         return true;
     }
